+++ src/features/Treege/Treege.tsx
@@ -1,5 +1,5 @@
 import axios from "axios";
-import { useEffect } from "react";
+import { useLayoutEffect } from "react";
 import { QueryClient, QueryClientProvider } from "react-query";
 import DarkTheme from "@/components/Theme/DarkTheme/DarkTheme";
 import AuthProvider from "@/context/Auth/AuthProvider";
@@ -9,16 +9,24 @@ import TreegeProvider from "@/features/Treege/context/TreegeProvider";
 import type { TreeNode } from "@/features/Treege/type/TreeNode";
 import "@/config/i18n";
 
-interface TreegeProps {
-  initialTree?: TreeNode;
-  authToken?: string;
-  endPoint?: string;
-}
+type TreegeProps =
+  | {
+      authToken?: string;
+      endPoint?: string;
+      initialTree?: never;
+      initialTreeId?: never;
+    }
+  | {
+      authToken?: string;
+      endPoint?: string;
+      initialTree: TreeNode;
+      initialTreeId: string;
+    };
 
 const queryClient = new QueryClient({ defaultOptions: { queries: { retry: false } } });
 
-const Treege = ({ authToken, endPoint, initialTree }: TreegeProps) => {
-  useEffect(() => {
+const Treege = ({ authToken, endPoint, initialTree, initialTreeId }: TreegeProps) => {
+  useLayoutEffect(() => {
     axios.defaults.baseURL = endPoint;
     axios.defaults.headers.common.Authorization = `Bearer ${authToken}`;
   }, [endPoint, authToken]);
@@ -26,7 +34,7 @@ const Treege = ({ authToken, endPoint, initialTree }: TreegeProps) => {
   return (
     <QueryClientProvider client={queryClient}>
       <AuthProvider authToken={authToken}>
-        <TreegeProvider endPoint={endPoint} initialTree={initialTree}>
+        <TreegeProvider endPoint={endPoint} initialTree={initialTree} initialTreeId={initialTreeId}>
           <DarkTheme>
             <SnackbarProvider>
               <TreeGrid />