+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # [Versions](https://github.com/Tracktor/treege/releases)
 
-## v0.4.20
-- Add repeatable option
+## v0.4.21
+- Update Treege props type
+- Remove helper text name to tree name text field
+- Default tree name props
+- Fix translation